requirements.txt
fastapi
uvicorn[standard]
fastapi-cache2[redis]
